pyproject.toml
[project]
name = "mcpsrv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "lxml>=6.0.0",
    "mcp[cli]>=1.12.2",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
]
